requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
Flask-Mail
line-bot-sdk
psycopg2-binary
python-dotenv
opencv-python-headless
requests
gunicorn
google-generativeai>=0.8.3
